+++ packages/charts/react-charting/src/types/IDataPoint.ts
@@ -608,6 +608,7 @@ interface ISLinkExtra {
    */
   value: number;
   unnormalizedValue?: number;
+  color?: string;
 }
 
 export type SNode = SankeyNode<ISNodeExtra, ISLinkExtra>;