+++ packages/react-monaco-editor/CHANGELOG.md
@@ -1,8 +1,17 @@
 # Change Log - @fluentui/react-monaco-editor
 
-This log was last generated on Wed, 10 Dec 2025 06:28:00 GMT and should not be manually modified.
+This log was last generated on Mon, 05 Jan 2026 14:47:18 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [1.8.5](https://github.com/microsoft/fluentui/tree/@fluentui/react-monaco-editor_v1.8.5)
+
+Mon, 05 Jan 2026 14:47:18 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-monaco-editor_v1.8.4..@fluentui/react-monaco-editor_v1.8.5)
+
+### Patches
+
+- Bump @fluentui/react-charting to v5.25.5 ([PR #35615](https://github.com/microsoft/fluentui/pull/35615) by beachball)
 
 ## [1.8.4](https://github.com/microsoft/fluentui/tree/@fluentui/react-monaco-editor_v1.8.4)